replace direct core lib usage with SciChartReact in TenorCurve3DChart component
